refactor!: modularize helpers
This commit removes the `conditional` helper used to register one of
several sources. You can replicate the same behavior by simply using a
function that checks a given condition and returns the source you want
to register.

`with({ condition = ... })` is not affected by this commit.
